requirements.txt
streamlit
langchain-google-genai
langchain
langchain-community
chromadb
python-dotenv
pypdf
beautifulsoup4
requests
langchain-text-splitters
